% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{activate_console}
\alias{activate_console}
\title{Activate Console Pane}
\usage{
activate_console()
}
\value{
Called for side effects; returns NULL invisibly.
}
\description{
Brings the console to the front and moves cursor focus to the console.
}
\examples{
\dontrun{
activate_console()
}
}
