% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{activate_source_editor}
\alias{activate_source_editor}
\title{Activate Source Editor}
\usage{
activate_source_editor()
}
\value{
Called for side effects; returns NULL invisibly.
}
\description{
Moves cursor focus to the source editor. If multiple source columns exist,
this will focus the most recently active editor.
}
\examples{
\dontrun{
activate_source_editor()
}
}
