% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{activate_terminal}
\alias{activate_terminal}
\title{Activate Terminal Pane}
\usage{
activate_terminal()
}
\value{
Called for side effects; returns NULL invisibly.
}
\description{
Creates a terminal if none exists, makes it visible, and moves cursor focus
to the terminal.
}
\examples{
\dontrun{
activate_terminal()
}
}
\concept{rstudio-control}
\concept{rstudio-shortcuts}
