% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{build_package}
\alias{build_package}
\title{Build Package}
\usage{
build_package(...)
}
\arguments{
\item{...}{Arguments passed to devtools::build()}
}
\value{
Returns the path to the built package invisibly.
}
\description{
Builds the package using devtools::build().
}
\examples{
\dontrun{
build_package()
}
}
