% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_fun.R
\name{comb_fun}
\alias{comb_fun}
\title{Formulas for biomarker combination}
\usage{
comb_fun(n, list_vars, target)
}
\arguments{
\item{target}{}
}
\description{
Formulas for biomarker combination
}
\examples{
set.seed(15272)
parallel::makeCluster(parallel::detectCores())
suppressPackageStartupMessages(library(ggplot2))
vars <- c("MUC16_2","ENRAGE","tPA","KLRD1_2","FASLG_2","FCRL6","LTBP2","OPG_2","LAG3_2" ,'CDH17',"CEACAM5","TLR3","TLR3","NDRG1","S100A11","CDH17","VIM","GBP2","ESM1","PRCP")
N <- list(1,2,3,4,5)
target = "y"
COMB_L <- comb_fun(n = N, list_vars = vars, target = "y")
}
