% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{database}
\alias{database}
\alias{dbc}
\alias{dbd}
\alias{ld_odbc}
\title{Database Connection}
\usage{
dbc(cfg = NULL, dbin = NULL, conn_string = NULL)

dbd(cn)

ld_odbc(cfg, dbin)
}
\arguments{
\item{cfg}{e.g. prod, devel}

\item{dbin}{database name}

\item{conn_string}{Optional. Given priority when provided}

\item{cn}{connection object to close}
}
\value{
connection
}
\description{
Database Connection
}
\section{Functions}{
\itemize{
\item \code{dbc()}: connect

\item \code{dbd()}: disconnect

\item \code{ld_odbc()}: TBD

}}
