% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQueryFetch.R
\name{database_utils}
\alias{database_utils}
\alias{dbQueryFetch}
\title{Fetch Query Data}
\usage{
dbQueryFetch(cn, qry, n = NULL)
}
\arguments{
\item{cn}{connection object}

\item{qry}{query to iterate and fetch until completion}

\item{n}{total rows to get in each iteration. If NULL (default) then value to be used is 1M}
}
\value{
Object of class data.table
}
\description{
Send Query and Fetch Until Complete
}
\section{Functions}{
\itemize{
\item \code{dbQueryFetch()}: internal function to iterate a query over a connection

}}
