% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_fun.R
\name{DEG_multi_contrast}
\alias{DEG_multi_contrast}
\title{Multiple contrast DEG}
\usage{
DEG_multi_contrast(dds, list_contrast, padj = 0.05, lfc.cutoff)
}
\arguments{
\item{lfc.cutoff}{}
}
\description{
Multiple contrast DEG
}
\examples{
contrast = c("AD_severity_scorad",  "mild", "healthy")
contrast1 = c("AD_severity_scorad",  "moderate", "healthy")
contrast2 = c("AD_severity_scorad",  "severe", "healthy")
contrast3 = c("AD_severity_scorad",  "moderate", "mild")
contrast4 = c("AD_severity_scorad",  "severe", "mild")
contrast5 = c("AD_severity_scorad",  "severe", "moderate")
list_contrast <- list("contrast", "contrast1", "contrast2", "contrast3", "contrast4", "contrast5")
res_tot <-DEG_multi_contrast(dds, list_contrast = list_contrast, lfc.cutoff = log2(1.5))
res_tot \%>\% arrange(expression) \%>\%
mutate_if(is.numeric, round, 3) \%>\%
tableHTML(width = c(30,rep(100, dim(res_tot)[2]))) \%>\%
  add_theme('rshiny-blue') \%>\%
 add_theme('colorize', color = 'turquoise4') \%>\%
 write_tableHTML(file = 'RNA_DEGs_all.html')

}
