% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{document_package}
\alias{document_package}
\title{Document Package}
\usage{
document_package(...)
}
\arguments{
\item{...}{Arguments passed to devtools::document()}
}
\value{
Returns the result from devtools::document() invisibly.
}
\description{
Runs devtools::document() to generate documentation.
}
\examples{
\dontrun{
document_package()
}
}
\concept{development}
