% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catprogress.R
\name{.catprogress}
\alias{.catprogress}
\title{Internal Function for Data.Tables}
\usage{
.catprogress(n)
}
\arguments{
\item{n}{print every n iterations}
}
\description{
Print progress of group based calculations when the by column is specified
}
