% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df.R
\name{extract_df}
\alias{extract_df}
\title{Extract columns containing the parterns}
\usage{
extract_df(DF, list_columns, keywords)
}
\arguments{
\item{keywords}{}
}
\description{
Extract columns containing the parterns
}
