% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcalogger-package.R
\docType{package}
\name{hcatools-package}
\alias{hcatools}
\alias{hcatools-package}
\title{hcatools: DB Connection & Logging}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Minimal package containing common hca tools used internally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/happycabbage/hcatools}
  \item Report bugs at \url{https://github.com/happycabbage/hcatools/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{no-reply@happycabbage.io}

}
\keyword{internal}
