% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{layout_four_column}
\alias{layout_four_column}
\title{Switch to 4-Column Layout}
\usage{
layout_four_column()
}
\value{
Called for side effects; returns NULL invisibly.
}
\description{
Shifts to a 4-column layout by adding two additional source columns to the
left side of the default 2-column layout. Note: This requires RStudio
configuration to support multiple source columns.
}
\examples{
\dontrun{
layout_four_column()
}
}
\concept{rstudio-control}
\concept{rstudio-shortcuts}
