% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{layout_two_column}
\alias{layout_two_column}
\title{Switch to Default 2-Column Layout}
\usage{
layout_two_column()
}
\value{
Called for side effects; returns NULL invisibly.
}
\description{
Sets RStudio to the default 2-column layout: top-left is source,
bottom-left is console, top-right contains environment and related panes,
bottom-right is the files panel and related panes.
}
\examples{
\dontrun{
layout_two_column()
}
}
\concept{rstudio-control}
\concept{rstudio-shortcuts}
