% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{load_all_code}
\alias{load_all_code}
\title{Load All Package Code}
\usage{
load_all_code(...)
}
\arguments{
\item{...}{Arguments passed to devtools::load_all()}
}
\value{
Returns the result from devtools::load_all() invisibly.
}
\description{
Loads all package code using devtools::load_all().
}
\examples{
\dontrun{
load_all_code()
}
}
\concept{development}
\concept{testing}
