% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{pkg_coverage}
\alias{pkg_coverage}
\title{Run Package Coverage}
\usage{
pkg_coverage()
}
\value{
Returns the coverage result object invisibly.
}
\description{
A convenience function to run coverage report on the current package.
}
\examples{
\dontrun{
pkg_coverage()
}
}
