% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtools-package.R
\docType{package}
\name{rdstools-package}
\alias{rdstools}
\alias{rdstools-package}
\title{rdstools: DB Connection & Logging}
\description{
Minimal package containing common tools used internally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/rdstools}
  \item Report bugs at \url{https://github.com/r-data-science/rdstools/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
