% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtools-package.R
\docType{package}
\name{rdtools-package}
\alias{rdtools}
\alias{rdtools-package}
\title{rdtools: DB Connection & Logging}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Minimal package containing common tools used internally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/rdtools}
  \item Report bugs at \url{https://github.com/r-data-science/rdtools/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
