% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_f.R
\name{return_f}
\alias{return_f}
\title{Replace strings containing patterns with new strings}
\usage{
return_f(key_word, replace_key)
}
\arguments{
\item{replace_key}{}
}
\description{
Replace strings containing patterns with new strings
}
