% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveHexImage.R
\name{saveHexImage}
\alias{saveHexImage}
\title{Generate Hex Image}
\usage{
saveHexImage(prompt)
}
\arguments{
\item{prompt}{A prompt for image generation}
}
\value{
Saves an image to the local working directory and returns the path invisibly
}
\description{
Generate Hex Image
}
