% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{singAUC}
\alias{singAUC}
\title{AUC for single biomarker}
\usage{
singAUC(X, y, plotROC = FALSE, alg = c("Wilcoxon", "ROC"))
}
\arguments{
\item{X}{numeric biomarkers}

\item{y}{vector of outcome}

\item{alg}{}
}
\description{
AUC for single biomarker
}
