% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{switch_theme_dark}
\alias{switch_theme_dark}
\title{Switch to Random Dark Theme}
\usage{
switch_theme_dark()
}
\value{
Called for side effects; returns NULL invisibly.
}
\description{
Convenience wrapper to apply a random dark theme.
}
\examples{
\dontrun{
switch_theme_dark()
}
}
\concept{rstudio-control}
\concept{rstudio-shortcuts}
