% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{switch_theme_light}
\alias{switch_theme_light}
\title{Switch to Random Light Theme}
\usage{
switch_theme_light()
}
\value{
Called for side effects; returns NULL invisibly.
}
\description{
Convenience wrapper to apply a random light theme.
}
\examples{
\dontrun{
switch_theme_light()
}
}
