% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableROC.R
\name{tableROC}
\alias{tableROC}
\title{Generate AUC table from matrix of biomarkers and outcome}
\usage{
tableROC(X, y, round.digit = 4, rank = FALSE)
}
\arguments{
\item{rank}{}
}
\description{
Generate AUC table from matrix of biomarkers and outcome
}
