% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{test_package}
\alias{test_package}
\title{Test Package}
\usage{
test_package(...)
}
\arguments{
\item{...}{Arguments passed to devtools::test()}
}
\value{
Returns the test results invisibly.
}
\description{
Runs package tests using devtools::test().
}
\examples{
\dontrun{
test_package()
}
}
\concept{development}
\concept{testing}
