% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcall.R
\name{catcall}
\alias{catcall}
\title{Print the code to construct data}
\usage{
catcall(vec)
}
\arguments{
\item{vec}{print the call for this data to the console}
}
\description{
Print the code to construct data
}
\examples{
# Print the code to recreate a vector
x <- 1:3
catcall(x)
}
