% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-utils.R
\name{dev-utils}
\alias{dev-utils}
\alias{report_coverage}
\title{Dev Tools}
\usage{
report_coverage(...)
}
\arguments{
\item{...}{arguments to covr::codecov}
}
\description{
Dev Tools
}
\section{Functions}{
\itemize{
\item \code{report_coverage()}: Run package coverage and upload report

}}
