% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_funs.R
\name{log_funs}
\alias{log_funs}
\alias{log_header}
\alias{open_log}
\alias{close_log}
\alias{..log}
\alias{log_ini}
\alias{log_end}
\alias{log_err}
\alias{log_wrn}
\alias{log_suc}
\alias{log_inf}
\alias{read_logs}
\title{Log Messages}
\usage{
log_header()

open_log(fnam = NULL, jobId = NULL)

close_log(gather = TRUE, ...)

..log(level = c("e", "i", "w", "s", "o", "c"), ...)

log_ini(msg = NULL, add = NULL)

log_end(msg = NULL, add = NULL)

log_err(msg = NULL, add = NULL)

log_wrn(msg = NULL, add = NULL)

log_suc(msg = NULL, add = NULL)

log_inf(msg = NULL, add = NULL)

read_logs(detail_parse = TRUE, detail_sep = "|", lf = NULL)
}
\arguments{
\item{fnam}{name for log file}

\item{jobId}{optional job number to group a set of log files}

\item{gather}{if TRUE, gather logs and return on close}

\item{...}{additional arguments to pass to read_logs}

\item{level}{param for internal function ..log}

\item{msg}{Main log message}

\item{add}{Additional details, typically a message from R}

\item{detail_parse}{if TRUE, will parse the details column in the logs}

\item{detail_sep}{Used to split the details column by separater (if detail_parse arg is TRUE)}

\item{lf}{log file name}
}
\description{
Log Messages
}
\section{Functions}{
\itemize{
\item \code{log_header()}: get log header

\item \code{open_log()}: open log

\item \code{close_log()}: close log

\item \code{..log()}: Internal function for logging

\item \code{log_ini()}: Log errors

\item \code{log_end()}: Log errors

\item \code{log_err()}: Log errors

\item \code{log_wrn()}: Log warning

\item \code{log_suc()}: Log success

\item \code{log_inf()}: Log info

\item \code{read_logs()}: Log errors

}}
\examples{
log_err(msg = "My error Message", add = "Error information from R")
log_err(msg = "My error Message")

}
