% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdstools-package.R
\docType{package}
\name{rdstools-package}
\alias{rdstools}
\alias{rdstools-package}
\title{rdstools: Logging & Dev Helpers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Minimal package containing common tools used for logging and development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/rdstools}
  \item Report bugs at \url{https://github.com/r-data-science/rdstools/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
