% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdstools-package.R
\docType{package}
\name{rdstools-package}
\alias{rdstools}
\alias{rdstools-package}
\title{rdstools: Data Science Development Tools for 'R'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utilities that support 'R' data-science project workflows by standardizing console logging, reproducible diagnostics, and package development tasks. The toolkit captures color-coded logs with persistent files, includes helpers to recreate data objects for debugging, and automates coverage reporting for interactive sessions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/rdstools}
  \item Report bugs at \url{https://github.com/r-data-science/rdstools/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
