% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-shortcuts.R
\name{rstudio-shortcuts}
\alias{rstudio-shortcuts}
\title{RStudio IDE Shortcuts}
\description{
A collection of functions to control RStudio IDE behavior, useful for
mapping to keyboard shortcuts or Stream Deck buttons via RStudio addins.
}
